Incident_IQ_Core/pyproject.toml
[project]
name = "incident-iq-core"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.6",
    "httpx>=0.28.1",
    "psycopg2>=2.9.11",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "sse-starlette>=3.2.0",
    "uvicorn>=0.40.0",
]
